+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0
 - Updating last bonding height for validators
+- Refactor Transactions in CMD
 
 ## 1.0.0
 - Refactoring Consensus module